Write generic functions to handle resolving individual flattened fields
